// cosmo-backend/dist/utils/logger.d.ts
export declare const logger: {
    info: (message: string, meta?: any) => void;
    error: (message: string, error?: any) => void;
    warn: (message: string, meta?: any) => void;
    debug: (message: string, meta?: any) => void;
    stream: {
        write: (message: string) => void;
    };
};
//# sourceMappingURL=logger.d.ts.map